requirements.txt
--find-links https://download.pytorch.org/whl/cu116/torch_stable

numpy>1.2
hydra-core==1.1.1
hydra-colorlog
matplotlib
opencv-python
omegaconf
plotly
ftfy
torch==2.2.2
torchsde
torchdiffeq
pytorch-lightning==2.0.8
gitpython
scipy
sentence-transformers
torchsde
einops
torchdiffeq
gym
moviepy==1.0.3
tqdm
termcolor
wandb
einops_exts
flash_attn==2.7.3
git+https://github.com/openai/CLIP.git
timm
